+++ wow-schema/src/main/kotlin/me/ahoo/wow/schema/WowModule.kt
@@ -23,6 +23,13 @@ import me.ahoo.wow.schema.kotlin.KotlinNullableCheck
 import me.ahoo.wow.schema.kotlin.KotlinReadOnlyCheck
 import me.ahoo.wow.schema.kotlin.KotlinRequiredCheck
 import me.ahoo.wow.schema.kotlin.KotlinWriteOnlyCheck
+import me.ahoo.wow.schema.typed.AggregateIdDefinitionProvider
+import me.ahoo.wow.schema.typed.CommandDefinitionProvider
+import me.ahoo.wow.schema.typed.DomainEventDefinitionProvider
+import me.ahoo.wow.schema.typed.DomainEventStreamDefinitionProvider
+import me.ahoo.wow.schema.typed.SnapshotDefinitionProvider
+import me.ahoo.wow.schema.typed.StateAggregateDefinitionProvider
+import me.ahoo.wow.schema.typed.StateEventDefinitionProvider
 
 class WowModule(private val options: Set<WowOption> = WowOption.ALL) : Module {
     override fun applyToConfigBuilder(builder: SchemaGeneratorConfigBuilder) {
