+++ wow-schema/src/main/kotlin/me/ahoo/wow/schema/typed/DomainEventStreamDefinitionProvider.kt
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-package me.ahoo.wow.schema
+package me.ahoo.wow.schema.typed
 
 import com.fasterxml.classmate.ResolvedType
 import com.github.victools.jsonschema.generator.CustomDefinition
